// src/app/features/dashboard/submodule/disabilities/models/incapacidad.model.ts
export interface Incapacidad {
  marcaTemporal: Date;
  Oficina: string;
  nombre_de_quien_recibio: string;
  Tipo_de_documento: string;
  Numero_de_documento: string;
  Temporal: string;
  numero_de_contrato: string;
  apellido: string;
  nombre: string;
  edad: number;
  sexo: string;
  empresa: string;
  centrodecosto: string;
  celular_o_telefono_01: string;
  celular_o_telefono_02: string;
  correoElectronico: string;
  fecha_de_ingreso_temporal: string;
  tipo_incapacidad: string;
  codigo_diagnostico: string;
  descripcion_diagnostico: string;
  F_inicio: string;
  dias_incapacidad: number;
  Dias_temporal: number;
  dias_eps: number;
  nombre_eps: string;
  fondo_de_pensiones: string;
  estado_incapacidad: string;
  prorroga: string;
  Incapacidad_transitada: string;
  numero_de_incapacidad: string;
  nit_de_la_IPS: string;
  ips_punto_de_atencion: string;
  observaciones: string;
  tipo_de_documento_doctor_atendido: string;
  numero_de_documento_doctor: string;
  nombre_doctor: string;
  estado_robot_doctor: string;
  link_incapacidad: string;
  historial_clinico?: string;
  dias_de_diferencia: number;
  Fecha_de_Envio_Incapacidad_Fisica: string;
  responsable_de_envio: string;
}

export interface IncapacidadResponse {
  incapacidades: Incapacidad[];
  reporte: {
    consecutivoSistema_id?: string;
    confirmacion_fecha_de_radicacion?: string;
    fecha_de_recepcion_de_la_incapacidad?: string;
    fecha_revision_por_parte_de_incapacidades?: string;
    estado_del_documento_incapacidad?: string;
    aquien_corresponde_el_pago?: string;
    fecha_de_radicado_eps?: string;
    numero_de_radicado?: string;
    a_donde_se_radico?: string;
    quien_radico?: string;
    respuesta_de_la_eps?: string;
    codigo_respuesta_eps?: string;
    fecha_de_respuesta_eps?: string;
    numero_de_incapacidad_eps?: string;
    dias_pagos_incapacidad?: string;
    valor_incapacidad?: string;
    numero_transaccion_eps_arl?: string;
    transaccion_empresa_usuaria?: string;
    quien_corresponde_el_pago_final?: string;
    respuesta_final_incapacidad?: string;
  };
}